tested new HltTCK
git-svn-id: http://svn.cern.ch/guest/lhcb/MooreOnline/trunk@169175 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 066b8a2faa874407e9d93a509d60fe8ae1a975af

+++ MooreOnline/MooreScripts/install/testinstall.sh
@@ -5,29 +5,35 @@ echo "============================================="
 
 export EXPECTV="HEAD"
 export EXPECTONV="v5r6"
+#options to use the nightly slots... need two of those options which work together
 export SPopts="--nightly lhcb-trigger-dev"
 export useSPopts=1
 echo "testing for MooreOnline "${EXPECTV}" with Online "${EXPECTONV}" and extra options: "${SPopts}
+#whether to delete existing directories or not
+export fresh=0
 
 startdir=`pwd`
 
 if [ "$TMPDIR" == "" ]; then export TMPDIR="/tmp/"${USER}; fi; if [ ! -d "$TMPDIR" ]; then mkdir "$TMPDIR"; fi; cd "$TMPDIR";
 
 
-if [ -d "satelite" ]; then
-  rm -rf satelite
-fi;
+if [ ${fresh} -eq 1 ]; then
+    if [ -d "satelite" ]; then
+	rm -rf satelite
+    fi;
 
-if [ -d "MOORE" ]; then
-  rm -rf MOORE
-fi;
+
+    if [ -d "MOORE" ]; then
+	rm -rf MOORE
+    fi;
+fi
 
 if [ ! -d "satelite" ]; then
-  mkdir satelite
+    mkdir satelite
 fi;
 
 if [ ! -d "MOORE" ]; then
-  mkdir MOORE
+    mkdir MOORE
 fi;
 
 cd satelite
@@ -44,9 +50,9 @@ echo "TEST SCRIPT: now trying to run install script"
 echo "============================================="
 
 a=""
-if [ $useSPopts -eq 0 ]; then
+if [ ${useSPopts} -eq 0 ]; then
   echo "Install command to be run"
-  echo "./InstallMoore.sh" ${EXPECTV} "--install-location" `pwd`"/MOORE" "--online-satelite" `pwd`"/satelite"
+  echo ${startdir}"/InstallMoore.sh" ${EXPECTV} "--install-location" `pwd`"/MOORE" "--online-satelite" `pwd`"/satelite"
   echo "---------------------------------------------"
 
   ${startdir}/InstallMoore.sh ${EXPECTV} --install-location `pwd`/MOORE --online-satelite `pwd`/satelite
@@ -65,4 +71,4 @@ echo "============================================="
 echo "TEST SCRIPT: now trying to run Moore test job"
 echo "============================================="
 cd MOORE
-(source ${TMPDIR}"/MOOREONLINE/MooreOnline_"${EXPECTV}"/InstallArea/x86_64-slc6-gcc48-opt/setupMoore.sh"; echo "TCK data at: "${HLTTCKROOT}; gaudirun.py ${MOOREROOT}"/tests/qmtest/moore.qms/physics.qms/2012.qmt" --option="from Configurables import Moore; Moore().EvtMax=1; Moore().OutputLevel=4;")
+(source ${TMPDIR}"/MOORE/MooreOnline_"${EXPECTV}"/InstallArea/x86_64-slc6-gcc48-opt/setupMoore.sh"; echo "TCK data at: "${HLTTCKROOT}; gaudirun.py ${MOOREROOT}"/tests/qmtest/moore.qms/physics.qms/2012.qmt" --option="from Configurables import Moore; Moore().EvtMax=1; Moore().OutputLevel=4;";)
